refactor(jadestern): 이메일 검증 함수 분리
